Pull efficiency log entries up one notch
