+++ internal/extender/resource.go
@@ -19,6 +19,7 @@ import (
 	"time"
 
 	demandapi "github.com/palantir/k8s-spark-scheduler-lib/pkg/apis/scaler/v1alpha2"
+	"github.com/palantir/k8s-spark-scheduler-lib/pkg/binpack"
 	"github.com/palantir/k8s-spark-scheduler-lib/pkg/resources"
 	"github.com/palantir/k8s-spark-scheduler/config"
 	"github.com/palantir/k8s-spark-scheduler/internal"
@@ -320,6 +321,7 @@ func (s *SparkSchedulerExtender) selectDriverNode(ctx context.Context, driver *v
 		driverNodeNames,
 		executorNodeNames,
 		availableNodesSchedulingMetadata)
+	efficiency := computeAvgPackingEfficiencyForResult(availableNodesSchedulingMetadata, packingResult)
 
 	svc1log.FromContext(ctx).Debug("binpacking result",
 		svc1log.SafeParam("availableNodesSchedulingMetadata", availableNodesSchedulingMetadata),
@@ -332,13 +334,17 @@ func (s *SparkSchedulerExtender) selectDriverNode(ctx context.Context, driver *v
 		svc1log.SafeParam("candidateExecutorNodes", executorNodeNames),
 		svc1log.SafeParam("driverNode", packingResult.DriverNode),
 		svc1log.SafeParam("executorNodes", packingResult.ExecutorNodes),
+		svc1log.SafeParam("avg packing efficiency CPU", efficiency.CPU),
+		svc1log.SafeParam("avg packing efficiency Memory", efficiency.Memory),
+		svc1log.SafeParam("avg packing efficiency GPU", efficiency.GPU),
+		svc1log.SafeParam("avg packing efficiency Max", efficiency.Max),
 		svc1log.SafeParam("binpacker", s.binpacker.Name))
 	if !packingResult.HasCapacity {
 		s.createDemandForApplicationInAnyZone(ctx, driver, applicationResources)
 		return "", failureFit, werror.Error("application does not fit to the cluster")
 	}
 
-	metrics.ReportPackingEfficiency(ctx, s.binpacker.Name, availableNodesSchedulingMetadata, packingResult)
+	metrics.ReportPackingEfficiency(ctx, s.binpacker.Name, efficiency)
 
 	s.removeDemandIfExists(ctx, driver)
 	metrics.ReportCrossZoneMetric(ctx, packingResult.DriverNode, packingResult.ExecutorNodes, availableNodes)
@@ -354,6 +360,17 @@ func (s *SparkSchedulerExtender) selectDriverNode(ctx context.Context, driver *v
 		return "", failureInternal, err
 	}
 	return packingResult.DriverNode, success, nil
+}
+
+func computeAvgPackingEfficiencyForResult(
+	nodesSchedulingMetadata resources.NodeGroupSchedulingMetadata,
+	packingResult *binpack.PackingResult) binpack.AvgPackingEfficiency {
+
+	packingEfficienciesDefault := make([]*binpack.PackingEfficiency, 0)
+	for _, packingEfficiency := range packingResult.PackingEfficiencies {
+		packingEfficienciesDefault = append(packingEfficienciesDefault, packingEfficiency)
+	}
+	return binpack.ComputeAvgPackingEfficiency(nodesSchedulingMetadata, packingEfficienciesDefault)
 }
 
 func (s *SparkSchedulerExtender) selectExecutorNode(ctx context.Context, executor *v1.Pod, nodeNames []string) (string, string, error) {
